requirements.txt
gradio>=4.0.0
modal>=0.60.0
weaviate-client>=4.0.0
langgraph>=0.0.15
python-dotenv>=1.0.0
pydantic>=2.0.0 
pydantic_settings